requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
astroid==3.3.6
async-timeout==5.0.1
attrs==24.2.0
bandit==1.8.0
black==24.10.0
blinker==1.9.0
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
ConfigArgParse==1.7
coverage==7.6.9
crashtest==0.4.1
dill==0.3.9
distlib==0.3.9
dulwich==0.21.7
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
flake8==7.1.1
flake8-docstrings==1.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
frozenlist==1.5.0
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.14.0
haversine==2.9.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
identify==2.6.3
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
installer==0.7.0
invoke==2.2.0
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
keyring==24.3.1
locust==2.32.4
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
mypy==1.13.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pbr==6.1.0
pep8-naming==0.14.1
pexpect==4.9.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
pre_commit==4.0.1
propcache==0.2.1
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.3.2
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-docker==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.36
starlette==0.41.3
stevedore==5.4.0
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==1.0.1
websockets==14.1
Werkzeug==3.1.3
xattr==1.1.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
